data/News_Scraping/requirements.txt
MySQL-python==1.2.5
SQLAlchemy==0.9.8
argparse==1.2.1
gevent==1.0.1
greenlet==0.4.5
python-dateutil==2.4.0
six==1.9.0
wsgiref==0.1.2
